advisor: Move common code for failed packages to the base class
This also fixes an issue with the Nexus IQ advisor using the wrong start
time in the summary.
